Split parser expression tests
