+++ tests/docs_truth/repo_hygiene/file_size.rs
@@ -10,6 +10,7 @@ mod integration_fixture_splits;
 mod lexer_and_monomorphize;
 mod module_system_splits;
 mod parser_enum_splits;
+mod parser_expression_splits;
 mod resolver_phase2_splits;
 mod resolver_validation_splits;
 mod resolver_validation_test_splits;
